Fixes #5, correct issue causing %change >128 to overflow

+++ application.py
@@ -587,7 +587,7 @@ def update_future_delta_percentiles(community, gcm, decade):
     # Handle NaN and NA values, then round to integer.
     dj["percent_change"] = (dj["delta"] / dj["events_ERA"]) * 100
     dj["percent_change"] = (
-        dj["percent_change"].replace([np.inf], 0).fillna(0).round(0).astype("int8")
+        dj["percent_change"].replace([np.inf], 0).fillna(0).round(0).astype("int")
     )
 
     # Group magnitude of changes into 5 quantiles
